Module samd_dma::storage

Storage for DMA system base and write-back addresses.

The DMA system requires two contiguous blocks of memory to hold the first descriptor and write-back location for each channel. This module provides StorageN structures to hold these blocks of memory, where 1 <= N <= 32.

Structs

Storage1
Storage type for base and write-back memory.

Storage2
Storage type for base and write-back memory.

Storage3
Storage type for base and write-back memory.

Storage4
Storage type for base and write-back memory.

Storage5
Storage type for base and write-back memory.

Storage6
Storage type for base and write-back memory.

Storage7
Storage type for base and write-back memory.

Storage8
Storage type for base and write-back memory.

Storage9
Storage type for base and write-back memory.

Storage10
Storage type for base and write-back memory.

Storage11
Storage type for base and write-back memory.

Storage12
Storage type for base and write-back memory.

UnsafeStorage
A user allocated storage type.

Traits

DmaStorage
Trait for accessing the base and write-back addresses of a storage unit.

Type Definitions

CHANMAX
The maximum amount of channels that can exist.